Begin encapsulating date generators and tools
For the cronjobs that send inventory notifications we’ll need queries that check specific date transition points in the data.

There are several bits of code that do this kind of work sprinkled through the code right now.

To keep things DRY, all these bits should be moved into a trait.

Search /src for ‘DateTime’ to find the fragments that need to be moved.

See the one fragment i’ve moved so far in the Trait and its usage in general code for a pattern to follow.